Create NativeMethods for PasswordManager.
